ChatSidebar: Refactor with latest Blue conventions
