+++ src/components/ChatSidebar/utils.js
@@ -1,0 +1,1 @@
+export const COMPONENT_KEY = 'ChatSidebar'
